* cplus-dem.c (main): Cast enum style to int.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@40242 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libiberty/cplus-dem.c
@@ -5148,7 +5148,7 @@ main (argc, argv)
 		skip_first = i;
 
 	      mbuffer[i] = 0;
-	      flags |= style;
+	      flags |= (int) style;
 	      result = cplus_demangle (mbuffer + skip_first, flags);
 	      if (result)
 		{
